# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Birrias API",
        "description": "API RESTful completa para gestión de torneos de fútbol amateur desarrollada con Laravel 12 + Breeze API + Sanctum",
        "contact": {
            "name": "Birrias API Support",
            "email": "admin@birrias.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Birrias API Server"
        }
    ],
    "paths": {
...
        "/api/tournaments": {
            "get": {
                "tags": [
                    "Tournaments"
                ],
                "summary": "Listar mis torneos",
                "description": "Obtiene la lista de torneos del usuario autenticado (solo admin)",
                "operationId": "57254b9990dcba8759536c0d4b057f57",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filtrar por estado del torneo",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "active",
                                "inactive"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de torneos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Sin permisos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Tournaments"
                ],
                "summary": "Crear nuevo torneo",
                "description": "Crea un nuevo torneo (solo admin)",
                "operationId": "6ba636b4dda0fb628ec1f7ce0f0c529c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "format"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Liga Birrias 2024"
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-03-01"
                                    },
                                    "inscription_fee_money": {
                                        "type": "number",
                                        "format": "decimal",
                                        "example": "150.00"
                                    },
                                    "currency": {
                                        "type": "string",
                                        "example": "USD"
                                    },
                                    "format": {
                                        "type": "string",
                                        "enum": [
                                            "league",
                                            "league_playoffs",
                                            "groups_knockout"
                                        ],
                                        "example": "league"
                                    },
                                    "groups": {
                                        "type": "integer",
                                        "example": "4"
                                    },
                                    "teams_per_group": {
                                        "type": "integer",
                                        "example": "4"
                                    },
                                    "playoff_size": {
                                        "type": "integer",
                                        "example": "8"
                                    },
                                    "rounds": {
                                        "type": "integer",
                                        "example": "2"
                                    },
                                    "home_away": {
                                        "type": "boolean",
                                        "example": "true"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Torneo creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ErrorResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Error message"
                    },
                    "errors": {
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "apiAuth": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}